+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased]
+
+### Features
+
+* Allow all users to create private channels [#3277](https://github.com/TryQuiet/quiet/issues/3277)
+
 ## [7.3.0]
 
 ### Features